irusb: uncppify: encapsulate dispatch_table definition /1
- encapsulate the current dispatch table type (std::map) in
  a typedef, move it to a C-compatible header 'irusb_dispatch.h'
- wrap init and free implementations in extern-c functions
